CMML3 merge2
