+++ HypoModPy/hypobase.py
@@ -5,6 +5,8 @@ import wx
 from pathlib import Path
 from pubsub import pub
 
+#from hypotools import *
+
 
 # Preference Flags
 basicmode = 0
@@ -34,6 +36,12 @@ class DiagEvent(wx.PyCommandEvent):
     def __init__(self, text=""):
         super().__init__(DiagEventType)
         self.text = text
+
+
+
+# def DiagWrite(text):
+#     #pub.sendMessage("diagbox", message=text)
+#     test = 1
 
 
 class TextFile():
