+++ HypoModPy/hypomain.py
@@ -1,17 +1,14 @@
 
-import os
+
+from HypoModPy.hypobase import *
+from HypoModPy.hypotools import *
+from HypoModPy.hypograph import *
+from HypoModPy.hyposcale import *
+from HypoModPy.hypoparams import *
+
+from spikemod import *
+
 import sys
-
-import wx
-from pubsub import pub
-
-from HypoModPy.hypobase import GetSystem
-from HypoModPy.hypotools import TextFile, ToolSet, DiagBox, ToolPanel, DiagWrite, SetDiagBoxTarget
-from HypoModPy.hypograph import GraphDisp, GraphPanel
-from HypoModPy.hyposcale import ScaleBox
-from HypoModPy.hypoparams import ParamCon, ID_ModBrowse
-from HypoModPy.hypodat import PlotDat
-
 
 NSApp = None
 NSApplication = None
@@ -34,6 +31,8 @@ class MainFrame(wx.Frame):
         super(MainFrame, self).__init__(None, wx.ID_ANY, title, pos, size)
         self.ostype = GetSystem()
         self.statusbar = self.CreateStatusBar()
+        #self.SetBackgroundColour(wx.WHITE)
+        #errorCode = ctypes.windll.shcore.SetProcessDpiAwareness(2)
 
         # Initialise ToolBoxes
         self.diagbox = DiagBox(self, "Diagnostic", wx.Point(0, 0), wx.Size(400, 500))
@@ -43,17 +42,12 @@ class MainFrame(wx.Frame):
         self.plotcon = None
         self.spikedatabox = None
 
-        # self.respath = rpath;  # defaults to "" for Windows, bundle resource path for OSX
-        # #self.mainpath = mpath
-        # self.app_path = os.getcwd()
-        # if self.ostype == "Windows": self.respath = self.app_path + "/HypoModPy/Resource"
-        # #else: self.respath = self.app_path + "/../HypoModPy/Resource". # depends on folder open in vscode
-        # else: self.respath = self.app_path + "/HypoModPy/Resource"
-
-        # self.respath = rpath
-
-        self.app_path = os.path.dirname(os.path.abspath(__file__))
-        self.respath = os.path.join(self.app_path, "Resource")
+        self.respath = rpath;  # defaults to "" for Windows, bundle resource path for OSX
+        #self.mainpath = mpath
+        self.app_path = os.getcwd()
+        if self.ostype == "Windows": self.respath = self.app_path + "/HypoModPy/Resource"
+        #else: self.respath = self.app_path + "/../HypoModPy/Resource". # depends on folder open in vscode
+        else: self.respath = self.app_path + "/HypoModPy/Resource"
 
         self.diagbox.Write("MainFrame app_path " + self.app_path + "\n")
         self.diagbox.Write("ostype " + self.ostype + "\nMainFrame respath " + self.respath + "\n")
@@ -220,6 +214,13 @@ class MainFrame(wx.Frame):
         event.Skip()
 
 
+    # def OnIconize(self, event):
+    #     iconized = event.IsIconized()
+    #     event.Skip()
+    #     if not iconized: wx.CallAfter(self.RestoreFromMinimize)
+    #     self.diagbox.Write('Icon Event\n')
+
+
     def OnIconize(self, event):
         iconized = event.IsIconized()
 
@@ -232,7 +233,21 @@ class MainFrame(wx.Frame):
                         tool.box.Show(True)
                         tool.box.SetPosition(self.GetPosition(), self.GetSize())
 
+        #event.Skip()
+
     
+    # def RestoreFromMinimize(self):
+    #     if self.IsIconized():
+    #         self.Restore()
+    #     self.Raise()
+    #     for tool in self.toolset.tools.values():
+    #         if tool.box and tool.visible:
+    #             tool.box.Show(True)
+    #             tool.box.SetPosition(self.GetPosition(), self.GetSize())
+    #     if hasattr(self, "systempanel") and self.systempanel and self.systempanel.IsShown():
+    #         self.systempanel.Raise()
+
+
 
 class SystemPanel(wx.Dialog):
     def __init__(self, mainwin, title):
@@ -347,18 +362,9 @@ class HypoMain(MainFrame):
             self.graphsizer.Add(graphpanel, 1, wx.EXPAND)
 
         # Mod Init
-        if modname == "Osmo":
-            from OsmoModPy.osmomod import OsmoMod
-            self.mod = OsmoMod(self, "osmomod")
-
-        if modname == "Spike":
-            from SpikeModPy.spikemod import SpikeMod
-            self.mod = SpikeMod(self, "spikemod")
-
-        if modname == "Agent":
-            from AgentModPy.agentmod import AgentMod
-            self.mod = AgentMod(self, "agentmod")
-
+        if modname == "Osmo": self.mod = OsmoMod(self, "osmomod")
+        if modname == "Spike": self.mod = SpikeMod(self, "spikemod")
+        if modname == "Agent": self.mod = AgentMod(self, "agentmod")
         self.mod.DefaultPlots()
         
         # Scale Box
