+++ HypoModPy/hypospikes.py
@@ -103,7 +103,7 @@ class SpikeDataPanel(ToolPanel):
         self.Layout()
 
 
-    def SetCount(self, count):
+    def SetDataCount(self, count):
        self.cellcount = count
        if self.cellindex > self.cellcount: self.cellindex = 0
        #neuropop.numneurons = count;
@@ -182,6 +182,7 @@ class SpikeDat():
 
         # initialise arrays for spike interval analysis
         self.histsize = 20000
+        #self.hist1 = pdata(self.histsize + 1)
         self.hist1 = pdata(self.histsize + 1)
         self.hist5 = pdata(self.histsize + 1)
         self.hist1norm = pdata(self.histsize + 1)
