--- hypodata.py
@@ -1,3 +1,0 @@
-
-
-
